+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [15.4.1](https://github.com/olimorris/codecompanion.nvim/compare/v15.4.0...v15.4.1) (2025-05-11)
+
+
+### Bug Fixes
+
+* **adapters:** stack overflow if default model is function ([#1417](https://github.com/olimorris/codecompanion.nvim/issues/1417)) ([2d28420](https://github.com/olimorris/codecompanion.nvim/commit/2d284203ec63136f72e75a61b1cb19abb31a1ebe)), closes [#1415](https://github.com/olimorris/codecompanion.nvim/issues/1415)
+
+## [15.4.0](https://github.com/olimorris/codecompanion.nvim/compare/v15.3.1...v15.4.0) (2025-05-10)
+
+
+### Features
+
+* **chat:** call slash commands with keymaps ([#1412](https://github.com/olimorris/codecompanion.nvim/issues/1412)) ([598af68](https://github.com/olimorris/codecompanion.nvim/commit/598af68b53394cfd43f7a098dcd463da86c5dc3f))
+
 ## [15.3.1](https://github.com/olimorris/codecompanion.nvim/compare/v15.3.0...v15.3.1) (2025-05-10)